+++ src/nl/hannahsten/texifyidea/inspections/latex/typesetting/LatexQuoteInspection.kt
@@ -10,13 +10,16 @@ import com.intellij.psi.util.PsiTreeUtil.findChildrenOfType
 import com.intellij.refactoring.suggested.extend
 import nl.hannahsten.texifyidea.inspections.InsightGroup
 import nl.hannahsten.texifyidea.inspections.TexifyInspectionBase
+import nl.hannahsten.texifyidea.lang.LatexPackage
+import nl.hannahsten.texifyidea.lang.commands.LatexGenericRegularCommand
 import nl.hannahsten.texifyidea.psi.LatexNormalText
 import nl.hannahsten.texifyidea.util.files.commandsInFileSet
 import nl.hannahsten.texifyidea.util.files.document
-import nl.hannahsten.texifyidea.util.parser.inMathContext
 import nl.hannahsten.texifyidea.util.magic.PatternMagic.quotePattern
-import nl.hannahsten.texifyidea.util.replaceString
+import nl.hannahsten.texifyidea.util.magic.cmd
+import nl.hannahsten.texifyidea.util.parser.inMathContext
 import nl.hannahsten.texifyidea.util.parser.requiredParameter
+import nl.hannahsten.texifyidea.util.replaceString
 import nl.hannahsten.texifyidea.util.toTextRange
 
 /**
@@ -65,6 +68,13 @@ class LatexQuoteInspection : TexifyInspectionBase() {
             "``" to "''",
         )
         val commands = file.commandsInFileSet()
+
+        // When Ligatures=TeXOff is set, straight quotes will be used. This setting can be applied in certain fontspec commands
+        val fontspecCommands = LatexGenericRegularCommand.entries.filter { it.dependency == LatexPackage.FONTSPEC }.map { it.cmd }
+        if (commands.filter { it.name in fontspecCommands }.any { it.text.contains("ligatures=TeXOff", ignoreCase = true) }) {
+            return emptyList()
+        }
+
         for (quoteMaker in quoteMakers) {
             for (command in commands) {
                 val opener = command.requiredParameter(quoteMaker.openParam)